add board and player classes
