+++ lib/player.rb
